Prep for Release
• Updated .gitignore for Build Folder
• Completed Setup Wizard
• Moved require Variables to `prevars.js`
• Fixed some Formatting
• Removed some unused code
• Moved App Functions to `functions.js`
• Updated App Reset
• Added popups, and toasts
• Udpated Styles

+++ src/scripts/script.js
@@ -1,5 +1,6 @@
 // On Load
 $(function () {
+    preload.checkEnvStatus();
     requirements();
     waitForRequirements();
     checkUpdates();
@@ -10,10 +11,51 @@ $(function () {
 // Core App Functions
 const core = {
     reset: () => {
-        localStorage.clear();
-        location.reload();
-    }
-}
+        popup
+            .fire({
+                title: 'Are you sure?',
+                html: `<p style="color: red; font-weight: bold;">An app reset will erase ALL DATA, are you sure you want to continue?</p>`,
+                icon: 'error',
+                showCancelButton: true,
+                confirmButtonColor: '#d33',
+                cancelButtonColor: '#3085d6',
+                confirmButtonText: 'Reset App (Erases All Data)',
+            })
+            .then((result) => {
+                if (result.isConfirmed) {
+                    localStorage.clear();
+                    popup
+                        .fire({
+                            title: 'App Reset',
+                            confirmButtonText: `Restart App`,
+                            html: `<span>Your app data has been deleted. Please restart the app to continue<span>`,
+                            icon: 'success',
+                        })
+                        .then(() => {
+                            location.reload();
+                        });
+                }
+            });
+    },
+
+    restart: (html, showCancelButton) => {
+        popup
+            .fire({
+                title: 'Restart App?',
+                html: `${html}`,
+                icon: 'warning',
+                showCancelButton: showCancelButton,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Restart App',
+            })
+            .then((result) => {
+                if (result.isConfirmed) {
+                    location.reload();
+                }
+            });
+    },
+};
 
 // Add Log Functions
 const log = {
@@ -53,12 +95,12 @@ const log = {
 
 const toast = {
     success: (msg) => {
-        notif.success({ message: `${msg}`, dismissible: true, });
+        notif.success({ message: `${msg}`, dismissible: true });
     },
     error: (msg) => {
-        notif.success({ message: `${msg}`, dismissible: true });
-    }
-}
+        notif.error({ message: `${msg}`, dismissible: true });
+    },
+};
 
 // Warn of Errors on Extended Loading Times
 setTimeout(() => {
@@ -68,12 +110,29 @@ setTimeout(() => {
 
     <ul>
         <li>Incorrect SnailyCAD Directory
-            <ul><li>(Installation or Pre-Existing Directory)</li></ul>
+            <ul>
+                <li>(Installation or Pre-Existing Directory)</li>
+            </ul>
         </li>
         <li>Error upon SnailyCAD Installation</li>
-        <li>Improperly Configured <code>.env</code> file.</li>
+        <li>Deleted/Moved SnailyCAD Directory</li>
     </ul>
     <hr>
     <span>You can <a href="#" onclick="core.reset();">reset</a> SnailyCAD Manager to repair errors.</span>
-    `)
+    `);
 }, 5000);
+
+const preload = {
+    checkEnvStatus: () => {
+        if (localStorage.envPending == `true`) {
+            log.add(`ENV is pending`, 1);
+            $(`env`).prepend(
+                `<p style="color: orange">You must fill out the ENV file for first run!</p>`
+            );
+            $(`#editEnv`).click();
+        }
+        if (!localStorage.envPending) {
+            log.add(`ENV is not pending`, 0);
+        }
+    },
+};
